Fixed a dependency issue with traverseHierarchy
